osa-differ: Refactor and filtered reports
This patch adds filtered report support to the osa-differ script and
refactors some code that was repeated more than once.

A user can provide ``--projects-only`` on the command line to see
only the commits that were made to OpenStack projects that OSA
builds. The user can also use ``--roles-only`` to limit the report
only to the changes in OSA roles.

All commits are rendered through the same jinja template, which
should make future cosmetic updates much easier. Each section for a
project or role now includes links back to the repository itself
and links to each commit.

The script now also shows the commits on OSA itself. That seems
fairly important and it should have been there earlier.

Change-Id: I139f0f1bab3d4afbc5004dc52756800e426aa605