Make osa-differ more friendly
This patch makes the --help output a bit more friendly for new users
and allows a user to provide a GitHub API token via an environment
variable to avoid GitHub's low API limits for unauthenticated users.

Logging is also available and users can enable debug logging to
get some reassurance that things are happening when the script
makes lots of requests to GitHub.

I'll follow this up with some documentation soon.

Change-Id: I17734f16dcecd2d62a202b99346d635fd0980856