Better handling of newton branch in osa-differ
There was a bug in osa-differ where an exception would be thrown
if a new role appeared that didn't exist before. That's fixed in
this patch.

The script also skips the commit comparison if it notices that the
commit SHA matches between both OSA releases.

The RST output had broken headers before and those are fixed as
well.

Closes-bug: 1604840

Change-Id: Ifa47d64e7b79639a90871fdbf0370f8c425582ed