Playbook to mount drives in swift object nodes
Currently in any of the release of openstack-ansible
there is no role to mount the drives in swift object
server nodes. This can now be a part of
openstack-ansible-ops as this is all operational stuff.

If the user knows the type of the file system that is
going to be used, we can mount the drives as well using
the playbook without manual intervention.

Doc is also modified according to the playbook.

Modified the lines according to the comments given.

Change-Id: I991ee4891253d2f99703c80e350df21c43db01e8
Closes-Bug: #1535536