Adds in ATFTPD role - Ubuntu 16.04
We have done an amount of work making the Multinode AIO work with Xenial
on both the deployment host and the booted virtual machines.  Sadly
Cobbler was a large complex system which required poking and prodding to
make it work properly on Xenial. Instead we have created a small system
which is a PXE environment with TFTPboot, DHCPD server and various
pre-seeds.

This work is virtually complete and working, this is a first commit of
many to get this merged in. Starting with a small commit to get one of
the few Ansible roles we have in place. After this has been checked over
and merged in, we will push all the rest of the work making sure any
amendments or recommendations are made to the rest of our work.

This first commit is simply an Ansible Role to install and configure
ATFTP on Ubuntu 16.04 deployment host, and does not currently affect
other items.

Change-Id: I9bfd88d81d27d5a6af7b297b9a4426f5e6200baf