Telegraf playbook fixes
The vm_quota.py is missing as telegraf plugin but still
referenced inside the playbook-influx-telegraf.yml
playbook.
Additionally the my.cnf is not necessary to be present
on the telegraf hosts/containers, in order to function.
The override influxdb_protocol exposes the protocol to
be used for communicating with influxdb, usually HTTP

Change-Id: I90226d02e82d2516be4a4d84baff22e46ce709fb