Make console logging optional
Console logging is useful when using the tool interactively, but
generats extra noise when using in a gating scenario. Move this into a
flag that enables it should it be necessary.

As part of this, the failure rate will be printed to stdout regardless
of console logging's status. This is so automated tools can use the
returned number to check for passing or failing.

Also, remove the old `configure_logging` function that was no longer in
use.

Change-Id: I55d410e6d504b03993f90cfd38dcd938390d02d5