Fix neutron agent container removal
This change fixes the removal of neutron agent containers so that it no
longer relies on the containers existing on the deploy host.

The file `leapfrog_remove_remaining_old_containers` is only created on
the deploy host, when combined with the fact that the original task
couldn't fail, it resulted in the removal of the containers silently
failing on multi-node builds.

This change gets the list of containers from the file on the deploy host
before trying to delete them on all hosts.

Change-Id: Ic95187fd7e7ff93c796ce01f296cb06a16ba72bd