Fix Newton cinder-manage check used in db-upgrades
cinder-manage service list does not output a usable, full, hostname;
just the short hostname is output.  To fix this we need to query the
database directly.

We also change the behavior of this operation to drop all services.  We
do this at the recommendation of upstream cinder.  This is safe because
all cinder services are stopped and the services will re-register when
restarted.

Related-Bug: #1712372
Change-Id: I6d845165ec22d4c2aeece0636a550e0b57050c22