Add pre- and post- stage hooks to leap redeploy process
It may be desired to execute additional playbooks during
the redeploy process.

This adds pre- and post- stage hooks before/after the
execution of the setup-hosts, setup-infrastructure and
setup-openstack stages. They will only be utilised
if the env vars are set to something. If they are
not then the leap process will continue as-is.

Change-Id: I60977648db784f038c67fe71ee6d2a01f4df7db2