Correct re-deploy hooks to allow unset variables
The patch https://review.openstack.org/498167 added
hooks which failed due to the '-u' set in the script
which does not tolerate unset variables.

The script has been modified to use parameter
expansion to allow each hook to be detected properly
as being set without failing due to unset variables.

Change-Id: Ic7dd889a2f00de8a59239abb37b2d451e854ef63