Wait for guest capabilities to appear
Deploying the multi-node-aio from master on a machine running Ubuntu
14.04 fails frequently as libvirt doesn't think it has the hvm
OS type.  I was able to manually run "virsh capabilities" shortly after
libvirt was installed and sure enough it didn't list any guest
capabilities.  Subsequent runs of "virsh capabilities" then returned
the <guest> XML element w/ <os_type>hvm</os_type> defined.

This commit simply adds a task that checks "virsh capabilities",
retrying up to 6 times if the <guest> element is not present.  From my
limit testing this seems sufficient to ensure that the domains are
defined and created successfully.

Lastly, we add a task to create /etc/libvirt/storage which is expected
to exist, but doesn't on a 14.04 deployment.

Change-Id: I158987270b71d3781e91d819fdcb02da736f3c1d