Add ironic image generation and node registration scripts
The image generation script will generate centos xenial and trusty
images along with the deploy image.  It will then upload the images into
glance and clean up after itself, other then the pacakges needed for
diskimage-builder and virtualenv this is all contained and easily
cleaned up.

The node registration script relies on the image generation script being
run first.  It will create flavors if needed, but the values for some of
the variables are designed to be modified (image_ram/cpu for example).

Change-Id: I247cc57f9ec1c961c302eac90be3e5ee2dfa7e2c