Enable support for hardware that brings the network up slowly
Add the slow-network element which defaults the timeout of network
bring-up to 60 seconds set by the NETWORK_TIMEOUT environment variable.

Syncs the ops repository to a known location so that we can use the
element when building images.

Rename the register-images script to register-nodes as that what it
actually does.

Change-Id: I403667f7bf6c69e255a4f5639ee4a8e0e8af18e3