Add env vars to set the RAM for VMS in the mnaio
The MNAIO serves as a good way to play with a production like OSA
environment on minimal hardware. While this is a good way to test it
does leave some options hard to address and change when trying to build
in an automated way. This change makes it possible for a deployer to set
an environmental variable to change the ram allocation of all default vm
types which should help folks build a more capable and customizable test
cloud.

Change-Id: Idc2cff091701ecda80165a98fffa1db222600d27
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>