Allow enabling of dhcp on provider network.
Dhcp is disabled by default for the provider network.  We are getting
complaints as teams using the mnaio env are expecting to be able to
attach devices to the provider network as well as the private network.
This change is to allow the setting of a 'enable_provider_net_dhcp' bool
variable to enable it during the service setup playbooks if needed.

Closes-Bug: 1755295

Change-Id: I46b22635fd8eece5528afecc76310f3c116b9e89