Allow users to skip services to cut down on conflicts
We have received complaints that the openstack-service-setup.yml
playbooks will try to overwrite flavors and images of the same
name that may have different attributes.  This is due to them
requiring some unique configurations that just so happen to
have been using the same name in tests.  The tag additions
will allow them to work around this by skipping the ones
conflicting with existing entries.

Change-Id: I3eef87f522a219d094181fa64933f86142b060ef
Closes-Bug: 1760219