mnaio service setup skips shade install if using tags instead of skip-tags
This was initially tested with using --skip-tags for the
create_network, create_images or create_flavors so you could
skip the ones you want. A task to install shade however is
skipped if someone uses --tags to specify a single tag. For
both to work, I'm adding an always tag to it.

Change-Id: Ia012a2417c8d6b10698ffe0c0712497eba8f4423
Closes-Bug: 1763080