Set threadpool tuning and fstype
The Elasticsearch cluster was running with the default thread pool, and
while functional, it will start rejecting work under heavy load. This
change sets the thread pool using the upstream guidelines for a more
highly performant cluster. The cluster is using niofs as the default
index store so that the kernel is better able to manage indexes instead
of only relying on memory mapped inodes.

Change-Id: I0239f1622c42cb25c21de69fd0ad5dc9e78ed6c5
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>