Add task to install requirements and change pathing
To run the curator playbook pip is required. This change installs the
package `python-virtualenv` to resolve the dependency and installs the
curator application into a virtualenv within /opt. This will keep the
host clean by isolating the dependencies and ensure that we can easily
clean up should we ever need to.

Change-Id: If872bb30281948d79a3f79a0e3a39e8a457c6fdb
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>