MNAIO: Add access to libvirt v4 via UCA Queens
When using an Ubuntu Xenial host, this patch adds the Ubuntu
Cloud Archive Queens repository so that the MNAIO tooling
makes use of libvirt v4. This provides access to a better
snapshot and snapshot-revert implementation, among other
features.

To improve the chances of success during builds, retries
are added to the package install tasks. Also, given that
we're using Ansible > 2.1.x, we forgo the with_items loop
for the package installs and just give the package module
the list so that it installs them all at once.

Change-Id: I0373e29fb996de1538465277760a0181289cbb44