Correct elasticsearch host roles for small clusters
Previously, the number of masters was equal to the node count divided
by two and rounded up if an even number. This worked for cluster sizes
greater than 3 nodes but did not properly assign enough masters for a
3 node cluster, and did not handle special cases of single node clusters
that may exist in a test environment.

New code handles these edge cases and improves the comments around node
role assignment.

Correcting the number of masters has the side effect of also correcting
the layout of data nodes. A 3 node cluster will no longer exhibit an
unused node.

Change-Id: I2243a6c0163ab32e3c0def383a10244a34450530