Add an 'always' tag to cluster layout task include
The templates which generate the configuration for the various
components require access to the facts defining the cluster layout.

Running the setup*.yml playbooks with --tags config would not include
common_task_data_node_hosts.yml and templating would fail due to
undefined vars.

This change applys an 'always' tag to all uses of the include.

Change-Id: I6c55a60aef54f9bb26eb9a2af08315265daa82d9