Improve metricbeat Ceph detection
When Ceph is deployed by OSA ceph_mons is a list of addresses
of the mons on the storage network populated from the OSA inventory.

When Ceph is deployed outside of OSA, ceph_mons is provided by
the deployer as a pointer to the mon addresses for integration.

To cover both these cases, the only OSA hosts in the control plane
which are guaranteed to be able to access the Ceph mons and the ELK
stack are those in the cinder_volumes host group.

This change configures metricbeat on the cinder_volumes hosts to
collect data from the ceph REST API on the mon hosts when possible.
Otherwise, collection is attempted from all hosts which have a
/etc/ceph directory.

Change-Id: Ib90f4c54636371b38d37e588ef6f7ba259fbe1b3