MNAIO: Run each setup playbook individually
The documentation describes using each setup-*.yml playbook
individually. This helps to reduce the memory usage on the
deploy host and also makes the output more readable as you're
only having to read the output of a single meta-playbook when
looking at the task results. In this patch we make the setup
process use the same set of playbooks described in the docs.

Change-Id: I596ed599de2e4302a82f2401f8fdf57f97660060