MNAIO: Disable SSH key checks on host
When accessing the VM's on an MNAIO host and doing multiple
rebuilds, the SSH keys are constantly changing. This creates
a situation where keys constantly have to be deleted and
accepted which isn't very user-friendly.

Given that this tooling is used for test purposes, we can
disable the host key checks without being too concerned.

Change-Id: I3dd1221c4789b0ab8e895b22b05906456fc1fc8f