Collect journals from containers as well as the host
When journal_paths is empty the function sdJournal.NewJournal()
is used to open the journal for the host system only.

A single entry in journal_paths is opened with
sdJournal.NewJournalFrom[Dir|Files](), and multiple entries are opened
only with sdJournnalFromFiles().

Adding a single entry of /var/log/journal in the config file causes
all journal files under that directory to be opened, rather than only
that of the host system.

Change-Id: Ib758407edebff6786bf64fcf95328fb89912e3f6