Enforce no_proxy when setting up ELK dashboards and rollups
There is no guarantee that all container IP addressess will be included
in an existing no_proxy environment variable. This will cause failures
when an http proxy is configured, but the proxy does not allow traffic
to 'hairpin' back to internal addresses.

This change forces no_proxy to the specific address of the kibana
and coordinator endpoints when the uri module is used to load dashboards
and configure rollups.

Change-Id: I669334c722cce79459b522e6e2d7e1aaec49ef24