MNAIO: Implement retries for image downloads
The image downloads may fail, even with aria's built-in
retry mechanism. With this patch we ensure that ansible
will delay and retry again. This improves the chances of
success.

With this we also remove the '--quiet' default parameter
so that we get console output from the task if it does
ultimately fail. This is useful for diagnostic purposes.

Change-Id: Ieed41f06a22effb28463637184980a748791edfe